ほぼstaticメソッドだったRecordクラスを削除してDatabaseクラスとしてDB処理をまとめた
